Use build id from Jenkins CI jobs in version number (#817)
This PR will use the build number of the job and include it into the version number. The version number will now also tell us when a build was compiled for internal or manually

Production builds will look like this:
2.15.2801

Internal builds will look like this:
2.15.2801-internal

Manually compiled builds will include the commit id (short SHA) of the last commit (the build number will be always 0 because no CI was used):
2.15.0-5df6b7c
or
2.15.0-5df6b7c-internal

To bump the major.minor part of the version number, you can use the grunt task:
grunt bump-version

In this case it would bump the version from 2.15 to 2.16.