fix: replace gpu-process-crashed with child-process-gone event handler
https://www.electronjs.org/docs/latest/api/app#event-gpu-process-crashed-deprecated
mentions the gpu-process-crashed event handler is deprecated:

> Deprecated: This event is superceded by the child-process-gone
> event which contains more information about why the child process
> disappeared.
> It isn't always because it crashed. The killed boolean can be replaced
> by checking reason === 'killed' when you switch to that event.

There's been some reports of people with AMD GPUs not beeing able to log
in with wire-desktop > 3.29.2997 (WEBAPP-7162).

Whenever the event occurs, the user is thrown back to the login screen,
but unable to login, as they're "already logged in".

This updates the handler to use child-process-gone, and only log event
and details, rather than invoking `main.reload()`, hoping to get some
insight into the kind of crash we're seeing here.

https://github.com/NixOS/nixpkgs/pull/196593#issuecomment-1300511999
https://github.com/NixOS/nixpkgs/issues/205707