Disable media session and media key features
When an audio message or a video is played, Wire is wrongly detected as
a media player and takes precedence over other music or video players
when a user presses hardware media keys.

The OS may also display Wire as a media player named "Chromium" with
play/pause, previous and next controls, of which only the play/pause
button is functional. It keeps being show even after the audio message
or the video has finished playing.